Add BotanicalBadAss patcher
The Botanical Bad A** achievement is broken in Miasmata, which I tracked
down to the game searching the wrong section of the save database for
the plant journal entries. They are under "plants", but the game was
searching under "plant".

There's enough padding after that string in the executable that it is
trivial to patch with a hex editor, and this program essentially does
exactly that, but gives users something they can just download and run
to fix the game.

I don't usually code in C#, so be nice :)