Add SMAP decoder
This script can decode the various SMAP type files in the game and
overlay them on the game map.

If you happen to be interested in the location of a particular (non
0xff) byte in the file it can also be used to highlight it's location.
Useful if you have determined what index is failing some comparison
using a debugger, e.g. for the Bored Cartographer achievement.