Add GUI SaveManager to check achievement progress
Easy way to:
- Check your progress mapping the coastline
- Check which Urns you have not lit yet
- Check how many notes & plants you have found
- Reset the Sanchez #1 not if it has glitched

Unfortunately it seems my use of the multiprocessing module does not mix
well with py2exe and it becomes unbearably slow. When not using py2exe
the speed is fine, so I'll either need to find a solution that preserves
the python interpreter or go back to my single threaded DDS decoder.