Fix off by one error generating alocalmod with no mods
This would use the wrong rs5 as the base if no mods were present on the
first run when alocalmod.rs5 was initially created.

If communitypatch.rs5 was installed it would use give a TypeError
(coercing to unicode) as it tried to use environment.rs5 as a base and
communitypatch.rs5 as though it was a .miasmod file.

If communitypatch.rs5 was not installed it would generate an attribute
error as it rolled off the top of the mod list.