Miaschiev: Add ability to reset head statues (ALL OF THEM)
Currently resets *EVERY* head statue (for testing) by removing the
exposed map around them. I need to test if that is sufficient - some of
the head statues may be referred to in the landmarks section of the save
game, which I have not yet reset (need to figure out how to work out
which landmark is which statue - the pos field is non-obvious).

If I confirm this is enough I'll just choose one of the statues on an
outlying island to reset, or ask the user.