Add Miasmata Map Fixes
These fix a couple of errors on the map - removed urns that don't exist
and fixed some cut off camp icons.

I actually made these last year, but haven't distributed them due to
their size and the fact that bundling them in anything other than
main.rs5 would cause Miasmata to crash on launch. Now that miaspatch is
on the way to address that issue it's time to finally release them ;-)