Add vi modelines to match my default style
My default style is really for the Linux Kernel, and I'm not fussed
about using it for python code, but most of my code here is already
using it, so I may as well stay with it - or at least be consistent.

After upgrading vim it seems to have started ignoring the settings from
my .vimrc for .py files, hence I need to set it explicitly now.