Add script to compose translated Journal index images
At the moment this only displays one line and cuts the text off -
wrapping the text AND centering it vertically is a bit of a problem.

I can set a fixed width to wrap the text easy enough, but I can only
guess as to what the height should be - too small and it cuts off, too
large and I won't be able to center it vertically.

Not sure what to do about this - maybe a patch to the GIMP to allow the
text box size to be set to fixed width & dynamic height? Or a patch to
explicitly support vertical centering? Or just hack together a word
wrapping algorithm in these scripts and leave the size as dynamic?