Add scripts to remove arbitrary text from textures
This script uses the heal selection plugin to remove dark black lines -
including most of the text. It leaves the image with three layers - a
hidden original (for comparison), a mask layer that I can manually
undelete any areas that are not text, and a layer with the heal
selection plugin applied.

The results will still need manual tweaking, but this will save me a
good deal of effort.