rs5-archive.py: Fix overwrite bug in add chunked file codepath
I didn't notice until now because the central directory archive in my main.rs5
was already at the end of the file from earlier modifications, so the file
pointer was already at EOF. The vanilla main.rs5 has the directory in the
middle, so this bug would overwrite arbitrary files. Nasty.