Procedure to wrap text around a mask
At the moment the mask is a merged version of the image - which works
for some images, but not other. However, I'll now now be able to easily
paint a mask in a layer of the image and use that instead, giving me a
huge amount of flexibility.

This could be quite useful in many situations and may well be worth
expanding and turning into a stand-alone GIMP plugin.