communitypatch/markers: Allow ruins near apple tree to be used for triangulation
This fix needs two parts - XL_LanmarkBox47 needs to be added to the
markers file with the same x,y,z, etc. as it's instance node, and
modelsets\XL_LanmarkBox47 needs to be added to marker_types in the
environment.

Unfortunately the markers file has to go in main.rs5 otherwise the game
will crash on launch, but I should have a solution to that in the near
future...