Generate mipmaps for certain textures
The original files all include mipmaps, but it works fine without them -
however some graphics cards may still benefit from having them in some
cases. I've picked a few where the textures are more likely to use them
- in world notes & items since they can be viewed from a distance, and
the map because it's so large.

I didn't bother with random menu and journal elements that are less
likely to benefit from mipmaps since they are never viewed from a
distance.

Also, it should be noted that Miaschiev relies on the map textures
including mipmaps to significantly speed up it's loading (currently it
fails if they don't, but that's a bug), so if nothing else they need
mipmaps for that.