Massive tweak to index texture font weight to match originals
I wasn't going to go to this length originally, but I want to correct an
error in the English version and don't want it to look out of place.

The bold setting in The GIMP/Pango is quite a bit heavier than the
originals, and trying to set a specific font weight doesn't seem to work
for whatever reason. I was able to use a convolution matrix to very
closely match the originals, though it mucked up for coloured text since
it didn't add up to 1 (and making it add up to 1 would necessarily
involve blurring it, which isn't what I'm after), so I had to do some
trickery duplicating, moving and re-combining the layers.