rs5-extractor: New ability to add undo metadata + revert file
This adds a new file inside the archive that contains details about what
the file used to look like - how large it was, and the location of the
original central directory. So long as the file is only modified by
appending data and appending new central directories it can later be
reverted back to the original state.

This is intended to be used by Miasmod when patching main.rs5 so that it
can be reverted back to the vanilla state if desired.