Disallow undo metadata to be repacked or manually added
This metadata will only make sense in the archive it was generated for.
Repacking the archive should discard it, and it should not be permitted
to add a file of it's name from elsewhere.

Still allow it to be extracted from an archive, which may be useful to
diagnose issues.