Add ability to add the contents of one rs5 into another
This will be used by Miasmod to handle patching main.rs5

The central directory of the mod being added is written to
MIASMOD\MODS\xxx.manifest *in addition* to the regular central
directory.

That will give me the ability to remove or update mods later on in a
sensible manner, particularly when mods may touch the same file -
consider for example that the journal index correction touches one of
the same files as the French translation, but clearly with the French
translation installed I would want that to override the English
correction, regardless of which order they are applied in.