Fix --list-chunks to work with the MIASMOD\ORIGDIR pseudo-file
ORIGDIR is not a real file - it's a placeholder over the location of the
original central directory to protect it from being inadvertently
overwritten (at this stage it is unnecessary when using this tool to
manipulate the rs5 files as the MIASMOD\UNDO file provides the necessary
protection for us, and if it proves problematic for Miasmata I'll simply
remove it).

The problem is, because it's not zlib compressed and doesn't look like
a typical file contained in an rs5 archive, anything that attempts to
read it as such will fail - this fixes one of those instances.