Add mod order support + significantly speed up hole search
The order of mods overriding each other can now be set inside
MIASMOD\ORDER and changed with an option to rs5-extractor.py

We no longer search for holes when adding every single file as this was
quite slow - instead, we now search for the largest hole in the file and
only search again if we were able to fill it with something. The hole
search algorithm is also speeded up significantly as it more quickly
dismisses anything before the undo metadata.

There is still room for improvement here - we could keep a list of holes
and keep it in sync so the search only has to be done once (this also
has the potential of later allowing us to optimise file placement to
reduce fragmentation), but that's a lower priority.