Add --list-holes option & remove ORIGDIR
This gives a detailed low level listing of all the location of all data
inside the archive and the unused space (holes) between them.

It includes the file header, central directory, the original central
directory if indicated by the undo metadata, files that have been masked
(replaced) by mods, and which mods have added which files.

I've decided to scrap the idea of MIASMOD\ORIGDIR for now - if someone
uses a tool that doesn't protect data before the undo metadata or that
has been masked by an installed mod they will corrupt things anyway. I
*could* bring this back by adding masked versions of files into the
central directory so that everything is protected...