communitypatch: Remove Glance Left/Right from settings menu
IonFX stated that this was removed at the last minute as it was causing
an issue with low-end systems, but they forgot to remove it from the
settings menu:

http://steamcommunity.com/app/223510/discussions/0/828925216501505816/#c828925216512703423

It might be nice to somehow add it back in, but for now let's just
remove the useless bindings from the settings menu. Also remove it from
the French translation (where I had previously just blanked out the
text).