Remove community patch assets from translation staging area
MiasPatch will be bundled with the community patch and translation
patch, so there won't be a need to stage the modified Map_FilledIn with
the translation files any more.

This will have the slight disadvantage of making the patch a bit larger
as it will contain two copies of Map_OverlayInfo (English & French), but
it still seems like the most sensible approach to me.