MiasPatch is ready :)
This all in one patcher can apply the Botanical Bad A** patch to
Miasmata.exe, install arbitrary environment mods and can even patch
main.rs5! From now on this will be the way I distribute updates to the
community patch, translation packs and any other mods that need to alter
files inside main.rs5.

It automatically detects any .miasmod and .rs5mod files that it was
bundled with and supports some limited configuration via miaspatch.cfg

When run, it scans the game directory to work out which bundled patches
need to be installed based on their version number so that it will do
something sensible if a newer version of one of the mods is already
installed - but the user can override that if necessary.

It also has the ability to completely remove all mods from the game -
removing the Botanical Bad A** fix, reverting main.rs5 and removing the
alocalmod.rs5 and the communitypatch.rs5 if they exist.

And given that this is intended to install language packs, the interface
is also fully translatable using Qt Linguist!