MiasPatch: Better handling of which mods to enable if no alocalmod.rs5
If alocalmod.rs5 doesn't exist we don't want to surprise the user by
installing more than we are saying we will (if it does exist we assume
that any enabled miasmod files are wanted), so disable any existing
installed mods.

alocalmod.miasmod is the one exception - that is always enabled
(otherwise Miasmod would get confused) and we preserve it if it exists
to avoid wiping out any local changes. Hopefully that won't trip anyone
up if they forgot they did something silly in it... We could ask the
user what to do with it, but I'm hoping to avoid MiasPatch becoming
overly complex - we have Miasmod for that.