Add initial support for extracting pk2 files from Obulis & Into Outland
These archives are used in IonFx' older games (I'm pretty sure this is
not the same .pk2 format used in Quake II). I noticed some similarities
between these and the rs5 format used in Miasmata (and plenty of
differences) and was interested to see how similar they and their
contained formats actually were (Obulis appears to use the same format
for sound files based on the common 0x00042A prefix, though I am yet to
work out what that is).

Unlike rs5 files these pk2 files have no file header or central
directory - each contained file is just placed one after another and has
to be read sequentially, much like a tar archive. It also does not use
any compression, but the local file headers and chunk structure is
similar to that used in rs5 files.

The unknown fields in the chunk headers contain different values than
those in rs5 files, so at the moment the code incorrectly dismisses them
as not being chunked files.  The u2 field in the local file headers
appear to be 0x30 for chunked files and 0x32 or 0x22 otherwise, so I
should be able to use that to determine if a file has chunks or not if I
wanted to continue with this line of investigation.