miasmod: If find_miasmata_save() fails, ask the user
The old code was a hack to allow me to test on Linux, but had the unfortunate
side effect that it would hide legitimate exceptions. This now asks the user
if it can't find the saves path automatically which should work for me on
Linux.

Turns out when I packaged up version 1.2 I did not have the winpaths module
installed on the box I was using (on account of my Alienware laptop where I
usually build this being dead) and the exception was masked so the button just
didn't seem to do anything.