rs5-extractor.py --analyse: Don't assert filename for MIASMOD\MODS\*
The mod info files are renamed when they are inserted in the rs5 file,
but I neglected to change the redundant filename embedded inside the
local file header as well. Not a huge deal, but remove the assertion for
these special cases since rs5 files now in the wild with this
inconsistency.