rs5-archive.py: Workaround weird timestamp bug in v2.0.0.4
ForgottenProdigy reported an issue where the menu in Miasmata was black
except for the created by message after using rs5-extractor.py.

This was tracked down to a strange issue in an old version (v2.0.0.4),
where it can fail depending on the timestamp on main.rs5

Running this before launching the game works:

    touch -d '16 Feb 2015 4:05AM' main.rs5

Yet running this fails:

    touch -d '16 Feb 2015 4:06AM' main.rs5

To work around the issue, force the timestamp to a known good one (needs
confirmation if this is timezone specific).