Patcher: 1st pass adding binary patcher to translate hardcoded strings
This scans the executable's read only data section for strings that can
be translated by matching on the context around them. Binary diffs are
not used as they tend to hide what is actually being modified and
require the executable to exactly match what they were created against,
but we want to support executables that may already have been modified
(e.g. by botanical.py), and may be any version from either Steam or GOG.

The lines to be translated are not matched to allow an already
translated executable to be matched and restored to English, or upgraded
to a new version of the translation, or converted from one language to
another.

Further, not all hunks have to be found to apply/remove the patch - this
allows the same patch to work with older versions of the game that did
not include all the settings, and increases the likelihood of the patch
working if Miasmata ever receives an update in the future.