Translation scripts: Simplify tag_preserving_split and fix edge cases
Rather than rejoining and splitting, rely on the fact that the original
partition inserts empty strings when a space was matched at the
start/end of a partition. If we join just the end of one partition with
the start of the next we will join any punctuation outside of a tag with
that tag, but still leave separate words split.

This removes the need for the tag_re_final_split regular expression and
eliminates the limitations imposed by it.