Spanish Translation: Notes, Maps, Misc
A few notes from David I need to take care of:

There are some notes with maps in them that contain references to
Outpost Sirius. Well, as you may know, Sirius is the name of a star,
like many other outpost names, and in Spanish it is so much better known
as "Sirio", so if you can, it would be great to change the name in notes
7, 8 and 11. I have created a TXT file for the note 11 with the
translated text for "to SIRIUS" that is included in the folder. Tell me
if you need me to make the other ones too, but you just need to write
"SIRIO" instead of "SIRIUS", cause there's no other word, unlike in note
11.

Sirius is the only star name that vary, the rest are left the same.

...

As for the Misc part, I added a TXT file for Lab_BlueBottle.