First pass of working randomizer
Plants and notes are shuffled around to create a new sort of challenge,
inspired by the Zelda randomizers.

Currently the types of plants are shuffled - meaning it will test your
knowledge of where all the different plants are located, but if you know
them all it is relatively straight forward.

Next will be changing it to shuffle around groups of plants instead,
meaning you would have to explore the entire island in search of the
cure.

Lots more testing required.