Randomizer: Output multi-coloured spoiler map
For testing I need to know where all the plants are, so output a spoiler
map with the various plants indicated. This is also intended to be an
option in the final randomizer to help stuck players. We might even look
at making this an overlay map with spoiler notes in game...

Brightest colours for the cure, half bright colours for other important
plants, dim colours for common plants. Tried to roughly match the colour
to the plants, while also trying not to have colours look too similar to
one another. Sorry colour blind players, for you we may need to output
individual spoiler maps for each plant.