Randomizer: Shuffle Fungi separately from other plants, add --seed
In testing I found if the Blue Fungi replaced the Rainbow Orchard it
would be below the ground and invisible to the player, and I saw other
instances of Fungi being hard to spot on the ground. It still could be
picked up by crouching and looking for the hand icon, but this isn't
something we want.

This moves all the Fungi that grow on the side of trees into a separate
shuffle bucket from the rest of the plants so their placement will be
reasonable.

This also adds a random --seed option so we can generate the same
randomizer again. Each shuffle bucket gets its own Random object so that
changing the shuffle mode of one bucket (e.g. notes) doesn't affect how
items in another bucket (e.g. plants) are shuffled.