Randomizer: Implement shuffle mode 3: clustering
This is the real meat and potatoes - this mode shuffles clusters of
plants around. That means part of the cure might be found where just a
single patch of previously insignificant wild flowers used to be, or
perhaps just a single yellow mushroom is now the blue fungus...

Welcome to Miasmata hard mode.